mutations: Float/Numeric arithmetic for increment/decrement
**Migration findings doc**: task 4, i106 (hecks-hecksagain-migration). Item 14 of the PR-split plan (corrected scope).

`CommandRules::Arithmetic#arithmetic`/`#arithmetic_value_object` were Integer-only. Widened to `Numeric` so Float fields (miette's own organ math, `increment: 0.02`) aren't rejected as a type mismatch.

**Real scope correction**: split out of the original plan's "item 14 — multiply + Float arithmetic support" bundle. Real-diff read of `6df3840` shows this widening touches only the two pre-existing ops' (`increment`/`decrement`) shared methods — `multiply` (item 15, next in the stack) is brand-new code with its own `Numeric` checks from birth and never calls `#arithmetic`/`#arithmetic_value_object` at all. So this is a genuinely separate finding, not entangled with `multiply` landing.

**What this PR does**
- Widens `#arithmetic`/`#arithmetic_value_object`'s type checks from `Integer` to `Numeric`.
- Adds `spec/mutation_float_arithmetic_growth_spec.rb`: increments and decrements a Float-typed value-object field.

**Verification**: `bundle exec rspec` — 1317 examples, 13 failures, same pre-existing failure set (tracked in `.pr-split-plan.md`; none are regressions). Confirmed via before/after: reverting this hunk raises `TypeMismatch` ("needs a value object with one shared Integer field") on both increment and decrement of a Float field.

Stacks after #46 (item 13, `remove`).

Fixes #149

+++ lib/hecksagain/runtime/command_rules/arithmetic.rb
@@ -55,11 +55,16 @@ module Hecksagain
             return arithmetic_value_object(current, amount, target, sign, op)
           end
 
-          unless amount.is_a?(Integer)
+          # Widened from Integer to Numeric (migration plan task 4, i106):
+          # miette's organ math increments a Float (`increment: 0.02`) --
+          # the raw, non-value-object path only ever mattered for Integer
+          # counters before this corpus existed. Integer stays the common
+          # case; Float is now accepted the same way.
+          unless amount.is_a?(Numeric)
             raise TypeMismatch, RefusalWording.render("TypeMismatch", "arithmetic_amount",
                                                        op: op, target: target, offered: Rendering.describe(amount))
           end
-          unless current.is_a?(Integer)
+          unless current.is_a?(Numeric)
             raise TypeMismatch, RefusalWording.render("TypeMismatch", "arithmetic_current",
                                                        op: op, target: target, offered: Rendering.describe(current))
           end
@@ -70,8 +75,14 @@ module Hecksagain
         def arithmetic_value_object(current, amount, target, sign, op)
           current_fields = current.to_h
           amount_fields  = amount.to_h
+          # Widened from Integer to Numeric -- see #arithmetic's own
+          # comment. A synthesised value-object wrapper around a bare
+          # Float attribute (miette's Synapse#strength, auto-wrapped per
+          # Part 3a's "bare primitives forbidden" finding) lands here as
+          # a one-Float-field Value exactly the way a one-Integer-field
+          # Value already did.
           shared_numeric = current_fields.keys.select do |field|
-            current_fields[field].is_a?(Integer) && amount_fields[field].is_a?(Integer)
+            current_fields[field].is_a?(Numeric) && amount_fields[field].is_a?(Numeric)
           end
           unless shared_numeric.size == 1
             raise TypeMismatch,
